chore(api): apply Rector findings + close Phase 8.1 #6,#7
The 2 Rector --dry-run findings from CI run 28968609054 are applied. The Rector CI step now exits 0 (clean) -- the formal all-gates-green signal (PHPStan green + PHPUnit green + Rector green) that closes Phase 8.1 candidates #6 and #7 from DOCS/PHASE_8_FOLLOWUPS.md and unlocks merging the PR.

Changes (5 in total across 2 files):

1. api/controllers/GoogleSheetsController.php:271 -- RecastingRemovalRector dropped the unnecessary (int) cast on curl_getinfo($ch, CURLINFO_HTTP_CODE). Downstream $httpCode >= 400 correctly handles the curl failure path (curl_getinfo returns false on failure -> false >= 400 evaluates to false). No semantic regression.

2. api/src/Controllers/JobSearchController.php lines 92, 93, 100, 101 -- RepeatedOrEqualToInArrayRector converted 4 duplicated `||` chains on $source to in_array($source, [...], true) form. Preserves strict (===) comparison semantics. Adzuna + Careerjet 2-way chains (lines 94, 95, 102, 103) intentionally left untouched -- Rector's rule fires only on chains with >= 3 comparisons, so applying the rule just to flagged sites avoids scope-creep into non-flagged chains that the user did NOT explicitly ask to convert.

Per the per-PR version rule, no version bump (CHANGELOG.md [2.6.35] ### Added (follow-up commits...) new bullet appended -- the 7th in that section).

+++ api/controllers/GoogleSheetsController.php
@@ -271,7 +271,7 @@ class GoogleSheetsController
         }
 
         $response = curl_exec($ch);
-        $httpCode = (int) curl_getinfo($ch, CURLINFO_HTTP_CODE);
+        $httpCode = curl_getinfo($ch, CURLINFO_HTTP_CODE);
         $error = curl_error($ch);
         curl_close($ch);
